+++ src/pages/index.tsx
@@ -19,7 +19,7 @@ export default function Home() {
       <Link href="/login" className="underline">
         登入
       </Link>
-      <Link href="/test/daliy" className="underline">
+      <Link href="/test/daily" className="underline">
         日常卡片測試
       </Link>
     </main>
