+++ src/pages/test/daily.tsx
@@ -1,9 +1,9 @@
-import Daliy from "@/components/recordCard/Daily";
+import Form from "@/containers/recordForm/daily/Form";
 
 export default function DaliyPage() {
   return (
     <main className="flex flex-col items-center justify-between p-24 min-h-screen bg-slate-400">
-      <Daliy />
+      <Form />
     </main>
   );
 }
